Finishing this program for now.
The program is in a state where all the core functionality works, and
now I just need to implement a real interface. I am going to leave this
for later, as the paper that this is meant for will be delayed for the
time being (SARS-COVID2).

What is left to do:

- Implement interface
   - CLI
   - Parsers
      - Tournament
      - Match data
- Write tests for the sampler, dataset, and summary classes.
- Find a dataset and run it, i.e. validation